// frontend/src/api/generated/types.gen.ts
export type GarmentOut = {
  id: number
  seller_id: number
  seller_name: string
  name: string
  description: string
  images: Array<string>
  price_plr: number
  size: string
  style: string
  condition: string
  location: string
  latitude: number | null
  longitude: number | null
  tags: Array<string>
  status: string
  created_at: string
}